extract the structured data stuff into a separate interface including a separate twig extension
