+++ src/Seo/SeoPage.php
@@ -16,7 +16,7 @@ namespace Sonata\SeoBundle\Seo;
 /**
  * http://en.wikipedia.org/wiki/Meta_element.
  */
-class SeoPage implements SeoPageInterface
+class SeoPage implements SeoPageInterface, StructuredDataAwarePage
 {
     /**
      * @var string
