+++ src/Twig/Extension/SeoExtension.php
@@ -257,20 +257,6 @@ class SeoExtension extends \Twig_Extension
     }
 
     /**
-     * Creates a script tag with type 'json-ld' and the JSON-LD string stored in page object.
-     *
-     * @return string
-     */
-    public function getStructuredData()
-    {
-        if (empty($this->page->getStructuredData())) {
-            return '';
-        }
-
-        return sprintf("<script type=\"application/ld+json\">%s</script>\n", $this->page->getStructuredData());
-    }
-
-    /**
      * @param string $string
      *
      * @return mixed
